Update FP_ITL_2s_l3cp_v28.seq
Replace 
SlowTimeP:    800000 ns
which will generate a time interval associated to a time slice larger than the limit  2^16  , by 
SlowTimeP:     400000 ns
and double the number of time slices in SlowFlush which use this parameter. Notice that SlowFlush will then reach the limit of 16 time slice for a function which is fine.